src/requirements.txt
fastapi
uvicorn
transformers
python-dotenv
librosa==0.10.2
python-multipart